[98] Move storages to `domain` package + improve DI
